requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
dataclasses-json==0.5.14
dateparser==1.1.1
defusedxml==0.7.1
docutils==0.20.1
fastapi==0.77.1
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
gpt4all==1.0.8
h11==0.14.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
joblib==1.3.1
khoj-assistant==0.10.0
langchain==0.0.247
langsmith==0.0.15
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.1.1
Pillow==9.3.0
pycparser==2.21
pydantic==1.10.12
Pygments==2.15.1
pypdf==3.14.0
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2022.3.2
requests==2.31.0
rich==13.5.0
safetensors==0.3.1
schedule==1.1.0
scikit-learn==1.3.0
scikit-surprise==1.1.3
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
shiboken6==6.5.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
starlette==0.19.1
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
uvicorn==0.17.6
yarl==1.9.2
